requirements.txt
tensorflow-gpu==1.15.2
mrcnn
h5py==2.10.0